Refactor: split app.rs and ui.rs into module trees
- app/ {mod,actions,commands,detail,mouse}: state+events, key actions,
  command prompt, detail view, mouse handling
- ui/ {mod,queue,cover,lyrics,detail,metadata,visualizer,footer,help}
- fix band wheel seek direction (up now rewinds)
- widen lyrics/metadata/detail scroll fields to usize (u16 truncation)
- clippy --fix pass (collapsible ifs, redundant casts/closures)

+++ src/mpd.rs
@@ -323,9 +323,9 @@ async fn run_command(client: &Client, command: MpdCommand) {
       .map_err(|error| error.to_string()),
     MpdCommand::NudgeSeek(delta) => {
       let mode = if delta >= 0 {
-        SeekMode::Forward(Duration::from_secs(delta.unsigned_abs() as u64))
+        SeekMode::Forward(Duration::from_secs(delta.unsigned_abs()))
       } else {
-        SeekMode::Backward(Duration::from_secs(delta.unsigned_abs() as u64))
+        SeekMode::Backward(Duration::from_secs(delta.unsigned_abs()))
       };
       client
         .command(Seek(mode))
